test: Correção dos testes e controllers existentes.

+++ src/main/java/com/vingcard/athos/interview/controller/LockController.java
@@ -1,10 +1,9 @@
 package com.vingcard.athos.interview.controller;
 
 import com.vingcard.athos.interview.persistence.entity.Lock;
-import com.vingcard.athos.interview.service.impl.LockServiceImpl;
+import com.vingcard.athos.interview.service.LockService;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,7 +14,7 @@ import java.util.List;
 @AllArgsConstructor
 public class LockController {
 
-	private final LockServiceImpl lockService;
+	private final LockService lockService;
 
 
 	/**
@@ -41,7 +40,11 @@ public class LockController {
 	 */
 	@GetMapping("/{serial}")
 	public ResponseEntity<Lock> getLockBySerial(@PathVariable String serial) {
-		return lockService.getLockBySerial(serial);
+		Lock lock = lockService.getLockBySerial(serial);
+		if (lock == null) {
+			return ResponseEntity.notFound().build();
+		}
+		return ResponseEntity.ok(lock);
 	}
 
 
@@ -64,12 +67,12 @@ public class LockController {
 	 * Path: /api/locks/{serial}
 	 * Method: PUT
 	 *
-	 * @param serial Lock Serial ID
+	 * @param serial      Lock Serial ID
 	 * @param lockDetails Lock Object
 	 * @return Lock newly updated
 	 */
 	@PutMapping("/{serial}")
-	public ResponseEntity<Lock> updateLock(@PathVariable String serial, @Valid @RequestBody Lock lockDetails) {
+	public Lock updateLock(@PathVariable String serial, @Valid @RequestBody Lock lockDetails) {
 		return lockService.updateLock(serial, lockDetails);
 	}
 
@@ -82,9 +85,13 @@ public class LockController {
 	 * @param serial Lock Serial ID
 	 * @return Return Status code 202 if success
 	 */
-	@ResponseStatus(HttpStatus.ACCEPTED)
 	@DeleteMapping("/{serial}")
 	public ResponseEntity<?> deleteLock(@PathVariable String serial) {
-		return lockService.deleteLock(serial);
+		if (!lockService.existsById(serial)) {
+			return ResponseEntity.notFound().build();
+		}
+
+		lockService.deleteLock(serial);
+		return ResponseEntity.ok().build();
 	}
 }
